+++ src/kls_region.h
@@ -158,8 +158,8 @@ void KLS_autoregion_on_temp_start(struct Koliseo_Temp* t_kls);
 void KLS_autoregion_on_temp_free(struct Koliseo_Temp* t_kls);
 void KLS_autoregion_on_temp_push(struct Koliseo_Temp* t_kls, ptrdiff_t padding, const char* caller, void* user);
 
-#undef KLS_DEFAULT_HOOKS
-#define KLS_DEFAULT_HOOKS (KLS_Hooks) { \
+#ifndef KLS_DEFAULT_HOOKS
+#define KLS_DEFAULT_HOOKS &(KLS_Hooks) { \
         .on_new_handler = &KLS_autoregion_on_new, \
         .on_free_handler = &KLS_autoregion_on_free, \
         .on_push_handler =  &KLS_autoregion_on_push, \
@@ -167,6 +167,15 @@ void KLS_autoregion_on_temp_push(struct Koliseo_Temp* t_kls, ptrdiff_t padding, 
         .on_temp_free_handler = &KLS_autoregion_on_temp_free, \
         .on_temp_push_handler = &KLS_autoregion_on_temp_push, \
     }
+#endif // KLS_DEFAULT_HOOKS
+
+#ifndef KLS_DEFAULT_EXTENSIONS_LEN
+#define KLS_DEFAULT_EXTENSIONS_LEN 1
+#endif // KLS_DEFAULT_EXTENSIONS_LEN
+
+#ifndef KLS_AUTOREGION_EXT_SLOT
+#define KLS_AUTOREGION_EXT_SLOT 0
+#endif // KLS_AUTOREGION_EXT_SLOT
 
 #include "koliseo.h"
 
@@ -263,3 +272,4 @@ void kls_usageReport_toFile(Koliseo *, FILE *);
 void kls_usageReport(Koliseo *);
 ptrdiff_t kls_type_usage(int, Koliseo *);
 ptrdiff_t kls_total_padding(Koliseo *);
+int kls_get_maxRegions_KLS_BASIC(Koliseo *kls);