+++ src/koliseo.c
@@ -333,9 +333,9 @@ void kls_log(Koliseo *kls, const char *tag, const char *format, ...)
  * @see kls_temp_end()
  */
 #ifndef KOLISEO_HAS_LOCATE
-Koliseo *kls_new_alloc_ext(ptrdiff_t size, kls_alloc_func alloc_func, kls_free_func free_func, KLS_Hooks ext_handlers, void* user)
+Koliseo *kls_new_alloc_ext(ptrdiff_t size, kls_alloc_func alloc_func, kls_free_func free_func, KLS_Hooks* ext_handlers, void** user, size_t ext_len)
 #else
-Koliseo *kls_new_alloc_ext_dbg(ptrdiff_t size, kls_alloc_func alloc_func, kls_free_func free_func, KLS_Hooks ext_handlers, void* user, Koliseo_Loc loc)
+Koliseo *kls_new_alloc_ext_dbg(ptrdiff_t size, kls_alloc_func alloc_func, kls_free_func free_func, KLS_Hooks ext_handlers, void** user, size_t ext_len, Koliseo_Loc loc)
 #endif // KOLISEO_HAS_LOCATE
 {
     if (size < (ptrdiff_t)sizeof(Koliseo)) {
@@ -367,8 +367,18 @@ Koliseo *kls_new_alloc_ext_dbg(ptrdiff_t size, kls_alloc_func alloc_func, kls_fr
         kls_set_conf(kls, KLS_DEFAULT_CONF);
         kls->stats = KLS_STATS_DEFAULT;
         kls->conf.kls_log_fp = stderr;
-        kls->hooks = ext_handlers;
-        kls->extension_data = user;
+        kls->hooks_len = ext_len;
+
+        for (size_t i=0; i < kls->hooks_len; i++) {
+            kls->hooks[i] = ext_handlers[i];
+        }
+        for (size_t i=0; i < kls->hooks_len; i++) {
+            if (user) {
+                kls->extension_data[i] = user[i];
+            } else {
+                kls->extension_data[i] = NULL;
+            }
+        }
         kls->free_func = free_func;
         kls->next = NULL;
 #ifdef KLS_DEBUG_CORE
@@ -379,9 +389,11 @@ Koliseo *kls_new_alloc_ext_dbg(ptrdiff_t size, kls_alloc_func alloc_func, kls_fr
                 kls + kls->offset);
 #endif
 
-        if (kls->hooks.on_new_handler != NULL) {
-            // Call on_new extension
-            kls->hooks.on_new_handler(kls);
+        for (size_t i=0; i < kls->hooks_len; i++) {
+            if (kls->hooks[i].on_new_handler != NULL) {
+                // Call on_new extension
+                kls->hooks[i].on_new_handler(kls);
+            }
         }
     } else {
 #ifndef KOLISEO_HAS_LOCATE
@@ -422,10 +434,27 @@ Koliseo *kls_new_alloc_dbg(ptrdiff_t size, kls_alloc_func alloc_func, kls_free_f
 #endif // KOLISEO_HAS_LOCATE
 {
 #ifndef KOLISEO_HAS_LOCATE
-    return kls_new_alloc_ext(size, alloc_func, free_func, KLS_DEFAULT_HOOKS, KLS_DEFAULT_EXTENSION_DATA);
+    return kls_new_alloc_ext(size, alloc_func, free_func, KLS_DEFAULT_HOOKS, KLS_DEFAULT_EXTENSION_DATA, KLS_DEFAULT_EXTENSIONS_LEN);
 #else
-    return kls_new_alloc_ext_dbg(size, alloc_func, free_func, KLS_DEFAULT_HOOKS, KLS_DEFAULT_EXTENSION_DATA, loc);
+    return kls_new_alloc_ext_dbg(size, alloc_func, free_func, KLS_DEFAULT_HOOKS, KLS_DEFAULT_EXTENSION_DATA, KLS_DEFAULT_EXTENSIONS_LEN, loc);
 #endif // KOLISEO_HAS_LOCATE
+}
+
+/**
+ * Takes a ptrdiff_t size.
+ * Allocates the backing memory for a Koliseo.
+ * Sets the KLS_Conf field to KLS_DEFAULT_CONF.
+ * Sets the fields with appropriate values if memory allocation was successful, goes to exit() otherwise.
+ * @param size The size for Koliseo data field.
+ * @return A pointer to the initialised Koliseo struct.
+ * @see Koliseo
+ * @see Koliseo_Temp
+ * @see KLS_DEFAULT_CONF
+ * @see kls_new_alloc()
+ */
+Koliseo *kls_new(ptrdiff_t size)
+{
+    return kls_new_alloc(size, KLS_DEFAULT_ALLOCF, KLS_DEFAULT_FREEF);
 }
 
 /**
@@ -444,9 +473,9 @@ Koliseo *kls_new_alloc_dbg(ptrdiff_t size, kls_alloc_func alloc_func, kls_free_f
  * @see kls_new_alloc_ext()
  * @see kls_set_conf()
  */
-Koliseo *kls_new_conf_alloc_ext(ptrdiff_t size, KLS_Conf conf, kls_alloc_func alloc_func, kls_free_func free_func, KLS_Hooks ext_handlers, void* user)
+Koliseo *kls_new_conf_alloc_ext(ptrdiff_t size, KLS_Conf conf, kls_alloc_func alloc_func, kls_free_func free_func, KLS_Hooks* ext_handlers, void** user, size_t ext_len)
 {
-    Koliseo *k = kls_new_alloc_ext(size, alloc_func, free_func, ext_handlers, user);
+    Koliseo *k = kls_new_alloc_ext(size, alloc_func, free_func, ext_handlers, user, ext_len);
     bool conf_res = kls_set_conf(k, conf);
     if (!conf_res) {
         fprintf(stderr,
@@ -475,7 +504,40 @@ Koliseo *kls_new_conf_alloc_ext(ptrdiff_t size, KLS_Conf conf, kls_alloc_func al
  */
 Koliseo *kls_new_conf_alloc(ptrdiff_t size, KLS_Conf conf, kls_alloc_func alloc_func, kls_free_func free_func)
 {
-    return kls_new_conf_alloc_ext(size, conf, alloc_func, free_func, KLS_DEFAULT_HOOKS, KLS_DEFAULT_EXTENSION_DATA);
+    return kls_new_conf_alloc_ext(size, conf, alloc_func, free_func, KLS_DEFAULT_HOOKS, KLS_DEFAULT_EXTENSION_DATA, KLS_DEFAULT_EXTENSIONS_LEN);
+}
+
+/**
+ * Takes a ptrdiff_t size and a KLS_Conf to configure the new Koliseo.
+ * Additional arguments are for extensions.
+ * Returns the new Koliseo.
+ * @param size The size for Koliseo data field.
+ * @param conf The KLS_Conf for the new Koliseo.
+ * @return A pointer to the initialised Koliseo struct, with wanted config.
+ * @see Koliseo
+ * @see KLS_Conf
+ * @see KLS_DEFAULT_CONF
+ * @see kls_new_alloc_ext()
+ */
+Koliseo *kls_new_conf_ext(ptrdiff_t size, KLS_Conf conf, KLS_Hooks* ext_handlers, void** user, size_t ext_len)
+{
+    return kls_new_conf_alloc_ext(size, conf, KLS_DEFAULT_ALLOCF, KLS_DEFAULT_FREEF, ext_handlers, user, ext_len);
+}
+
+/**
+ * Takes a ptrdiff_t size and a KLS_Conf to configure the new Koliseo.
+ * Returns the new Koliseo.
+ * @param size The size for Koliseo data field.
+ * @param conf The KLS_Conf for the new Koliseo.
+ * @return A pointer to the initialised Koliseo struct, with wanted config.
+ * @see Koliseo
+ * @see KLS_Conf
+ * @see KLS_DEFAULT_CONF
+ * @see kls_new_alloc_ext()
+ */
+Koliseo *kls_new_conf(ptrdiff_t size, KLS_Conf conf)
+{
+    return kls_new_conf_alloc(size, conf, KLS_DEFAULT_ALLOCF, KLS_DEFAULT_FREEF);
 }
 
 /**
@@ -493,7 +555,7 @@ Koliseo *kls_new_conf_alloc(ptrdiff_t size, KLS_Conf conf, kls_alloc_func alloc_
  * @see KLS_Conf
  * @see kls_new_conf_alloc_ext()
  */
-Koliseo *kls_new_traced_alloc_handled_ext(ptrdiff_t size, const char *output_path, kls_alloc_func alloc_func, kls_free_func free_func, KLS_Err_Handlers err_handlers, KLS_Hooks ext_handlers, void* user)
+Koliseo *kls_new_traced_alloc_handled_ext(ptrdiff_t size, const char *output_path, kls_alloc_func alloc_func, kls_free_func free_func, KLS_Err_Handlers err_handlers, KLS_Hooks* ext_handlers, void** user, size_t ext_len)
 {
 
 #ifndef KLS_DEBUG_CORE
@@ -512,7 +574,7 @@ Koliseo *kls_new_traced_alloc_handled_ext(ptrdiff_t size, const char *output_pat
                                    .err_handlers.PTRDIFF_MAX_handler = ( err_handlers.PTRDIFF_MAX_handler != NULL ? err_handlers.PTRDIFF_MAX_handler : &KLS_PTRDIFF_MAX_default_handler_dbg__),
 #endif // KOLISEO_HAS_LOCATE
     };
-    return kls_new_conf_alloc_ext(size, k, alloc_func, free_func, ext_handlers, user);
+    return kls_new_conf_alloc_ext(size, k, alloc_func, free_func, ext_handlers, user, ext_len);
 }
 
 /**
@@ -531,7 +593,7 @@ Koliseo *kls_new_traced_alloc_handled_ext(ptrdiff_t size, const char *output_pat
  */
 Koliseo *kls_new_traced_alloc_handled(ptrdiff_t size, const char *output_path, kls_alloc_func alloc_func, kls_free_func free_func, KLS_Err_Handlers err_handlers)
 {
-    return kls_new_traced_alloc_handled_ext(size, output_path, alloc_func, free_func, err_handlers, KLS_DEFAULT_HOOKS, KLS_DEFAULT_EXTENSION_DATA);
+    return kls_new_traced_alloc_handled_ext(size, output_path, alloc_func, free_func, err_handlers, KLS_DEFAULT_HOOKS, KLS_DEFAULT_EXTENSION_DATA, KLS_DEFAULT_EXTENSIONS_LEN);
 }
 
 /**
@@ -546,10 +608,10 @@ Koliseo *kls_new_traced_alloc_handled(ptrdiff_t size, const char *output_path, k
  * @see KLS_Conf
  * @see kls_new_conf_alloc()
  */
-Koliseo *kls_new_traced_ext(ptrdiff_t size, const char *output_path, KLS_Hooks ext_handlers, void* user)
+Koliseo *kls_new_traced_ext(ptrdiff_t size, const char *output_path, KLS_Hooks* ext_handlers, void** user, size_t ext_len)
 {
     KLS_Err_Handlers err_handlers = KLS_DEFAULT_ERR_HANDLERS;
-    return kls_new_traced_alloc_handled_ext(size, output_path, KLS_DEFAULT_ALLOCF, KLS_DEFAULT_FREEF, err_handlers, ext_handlers, user);
+    return kls_new_traced_alloc_handled_ext(size, output_path, KLS_DEFAULT_ALLOCF, KLS_DEFAULT_FREEF, err_handlers, ext_handlers, user, ext_len);
 }
 
 /**
@@ -572,6 +634,39 @@ Koliseo *kls_new_traced_alloc(ptrdiff_t size, const char *output_path, kls_alloc
 }
 
 /**
+ * Takes a ptrdiff_t size and a filepath for the trace output file.
+ * Returns a pointer to the prepared Koliseo.
+ * Calls kls_new_conf_alloc() to initialise the Koliseo with the proper config for a traced Koliseo, logging to the passed filepath.
+ * @param size The size for Koliseo data field.
+ * @param output_path The filepath for log output.
+ * @return A pointer to the initialised Koliseo struct, with wanted config.
+ * @see Koliseo
+ * @see KLS_Conf
+ * @see kls_new_conf_alloc()
+ */
+Koliseo *kls_new_traced(ptrdiff_t size, const char *output_path)
+{
+    return kls_new_traced_alloc(size, output_path, KLS_DEFAULT_ALLOCF, KLS_DEFAULT_FREEF);
+}
+
+/**
+ * Takes a ptrdiff_t size and a filepath for the trace output file.
+ * Returns a pointer to the prepared Koliseo.
+ * Calls kls_new_conf_alloc() to initialise the Koliseo with the proper config for a traced Koliseo, logging to the passed filepath.
+ * @param size The size for Koliseo data field.
+ * @param output_path The filepath for log output.
+ * @param err_handlers The error handlers struct for errors in push calls.
+ * @return A pointer to the initialised Koliseo struct, with wanted config.
+ * @see Koliseo
+ * @see KLS_Conf
+ * @see kls_new_conf_alloc()
+ */
+Koliseo *kls_new_traced_handled(ptrdiff_t size, const char *output_path, KLS_Err_Handlers err_handlers)
+{
+    return kls_new_traced_alloc_handled(size, output_path, KLS_DEFAULT_ALLOCF, KLS_DEFAULT_FREEF, err_handlers);
+}
+
+/**
  * Takes a ptrdiff_t size, an allocation function pointer and a free function pointer, and returns a pointer to the prepared Koliseo.
  * Additional arguments are for extensions.
  * Calls kls_new_conf_alloc() to initialise the Koliseo with the proper config for a debug Koliseo (printing to stderr).
@@ -584,7 +679,7 @@ Koliseo *kls_new_traced_alloc(ptrdiff_t size, const char *output_path, kls_alloc
  * @see KLS_Conf
  * @see kls_new_conf_alloc_ext()
  */
-Koliseo *kls_new_dbg_alloc_handled_ext(ptrdiff_t size, kls_alloc_func alloc_func, kls_free_func free_func, KLS_Err_Handlers err_handlers, KLS_Hooks ext_handlers, void* user)
+Koliseo *kls_new_dbg_alloc_handled_ext(ptrdiff_t size, kls_alloc_func alloc_func, kls_free_func free_func, KLS_Err_Handlers err_handlers, KLS_Hooks* ext_handlers, void** user, size_t ext_len)
 {
 #ifndef KLS_DEBUG_CORE
     fprintf(stderr,
@@ -601,7 +696,7 @@ Koliseo *kls_new_dbg_alloc_handled_ext(ptrdiff_t size, kls_alloc_func alloc_func
                                .err_handlers.PTRDIFF_MAX_handler = ( err_handlers.PTRDIFF_MAX_handler != NULL ? err_handlers.PTRDIFF_MAX_handler : &KLS_PTRDIFF_MAX_default_handler_dbg__),
 #endif // KOLIEO_HAS_LOCATE
     };
-    Koliseo * kls = kls_new_conf_alloc_ext(size, k, alloc_func, free_func, ext_handlers, user);
+    Koliseo * kls = kls_new_conf_alloc_ext(size, k, alloc_func, free_func, ext_handlers, user, ext_len);
     kls->conf.kls_verbose_lvl = 1;
     return kls;
 }
@@ -620,7 +715,7 @@ Koliseo *kls_new_dbg_alloc_handled_ext(ptrdiff_t size, kls_alloc_func alloc_func
  */
 Koliseo *kls_new_dbg_alloc_handled(ptrdiff_t size, kls_alloc_func alloc_func, kls_free_func free_func, KLS_Err_Handlers err_handlers)
 {
-    return kls_new_dbg_alloc_handled_ext(size, alloc_func, free_func, err_handlers, KLS_DEFAULT_HOOKS, KLS_DEFAULT_EXTENSION_DATA);
+    return kls_new_dbg_alloc_handled_ext(size, alloc_func, free_func, err_handlers, KLS_DEFAULT_HOOKS, KLS_DEFAULT_EXTENSION_DATA, KLS_DEFAULT_EXTENSIONS_LEN);
 }
 
 /**
@@ -633,10 +728,10 @@ Koliseo *kls_new_dbg_alloc_handled(ptrdiff_t size, kls_alloc_func alloc_func, kl
  * @see KLS_Conf
  * @see kls_new_conf()
  */
-Koliseo *kls_new_dbg_ext(ptrdiff_t size, KLS_Hooks ext_handlers, void* user)
+Koliseo *kls_new_dbg_ext(ptrdiff_t size, KLS_Hooks* ext_handlers, void** user, size_t ext_len)
 {
     KLS_Err_Handlers err_handlers = KLS_DEFAULT_ERR_HANDLERS;
-    return kls_new_dbg_alloc_handled_ext(size, KLS_DEFAULT_ALLOCF, KLS_DEFAULT_FREEF, err_handlers, ext_handlers, user);
+    return kls_new_dbg_alloc_handled_ext(size, KLS_DEFAULT_ALLOCF, KLS_DEFAULT_FREEF, err_handlers, ext_handlers, user, ext_len);
 }
 
 /**
@@ -654,6 +749,35 @@ Koliseo *kls_new_dbg_alloc(ptrdiff_t size, kls_alloc_func alloc_func, kls_free_f
 {
     KLS_Err_Handlers err_handlers = KLS_DEFAULT_ERR_HANDLERS;
     return kls_new_dbg_alloc_handled(size, alloc_func, free_func, err_handlers);
+}
+
+/**
+ * Takes a ptrdiff_t size and returns a pointer to the prepared Koliseo.
+ * Calls kls_new_conf_alloc() to initialise the Koliseo with the proper config for a debug Koliseo (printing to stderr).
+ * @param size The size for Koliseo data field.
+ * @return A pointer to the initialised Koliseo struct, with wanted config.
+ * @see Koliseo
+ * @see KLS_Conf
+ * @see kls_new_conf()
+ */
+Koliseo *kls_new_dbg(ptrdiff_t size)
+{
+    return kls_new_dbg_alloc(size, KLS_DEFAULT_ALLOCF, KLS_DEFAULT_FREEF);
+}
+
+/**
+ * Takes a ptrdiff_t size and returns a pointer to the prepared Koliseo.
+ * Calls kls_new_conf_alloc() to initialise the Koliseo with the proper config for a debug Koliseo (printing to stderr).
+ * @param size The size for Koliseo data field.
+ * @param err_handlers The error handlers for errors in push calls.
+ * @return A pointer to the initialised Koliseo struct, with wanted config.
+ * @see Koliseo
+ * @see KLS_Conf
+ * @see kls_new_conf()
+ */
+Koliseo *kls_new_dbg_handled(ptrdiff_t size, KLS_Err_Handlers err_handlers)
+{
+    return kls_new_dbg_alloc_handled(size, KLS_DEFAULT_ALLOCF, KLS_DEFAULT_FREEF, err_handlers);
 }
 
 /**
@@ -993,7 +1117,7 @@ int kls__check_available_failable_dbg(Koliseo* kls, ptrdiff_t size, ptrdiff_t al
                 return -1;
             }
         } else if (OOM_happened) {
-            if (current->conf.err_handlers.PTRDIFF_MAX_handler && OOM_handled) {
+            if (current->conf.err_handlers.OOM_handler && OOM_handled) {
 #ifndef KOLISEO_HAS_LOCATE
                 fprintf(stderr, "[KLS] %s(): OOM fault happened and was handled.\n", caller_name);
 #ifdef KLS_DEBUG_CORE
@@ -1022,7 +1146,7 @@ int kls__check_available_failable_dbg(Koliseo* kls, ptrdiff_t size, ptrdiff_t al
 bool kls__try_grow(Koliseo* kls, ptrdiff_t needed)
 {
     ptrdiff_t new_size = KLS_MAX(kls->size * 2, needed);
-    Koliseo* new_kls = kls_new_conf_alloc_ext(new_size, kls->conf, KLS_DEFAULT_ALLOCF, KLS_DEFAULT_FREEF, kls->hooks, kls->extension_data);
+    Koliseo* new_kls = kls_new_conf_alloc_ext(new_size, kls->conf, KLS_DEFAULT_ALLOCF, KLS_DEFAULT_FREEF, kls->hooks, kls->extension_data, kls->hooks_len);
     kls_log(kls, "DEBUG", "%s(): growing Koliseo, new size: {%td}", __func__, new_size);
     if (!new_kls) return false;
     kls->next = new_kls;
@@ -1263,25 +1387,27 @@ void *kls_push_zero_ext_dbg(Koliseo *kls, ptrdiff_t size, ptrdiff_t align,
     current->prev_offset = current->offset;
     current->offset += padding + size * count;
 
-    if (current->hooks.on_push_handler != NULL) {
-        /*
-        struct KLS_EXTENSION_AR_DEFAULT_ARGS {
-            const char* region_name;
-            size_t region_name_len;
-            const char* region_desc;
-            size_t region_desc_len;
-            int region_type;
-        };
-        struct KLS_EXTENSION_AR_DEFAULT_ARGS ar_args = {
-            .region_name = KOLISEO_DEFAULT_REGION_NAME,
-            .region_name_len = strlen(KOLISEO_DEFAULT_REGION_NAME),
-            .region_desc = KOLISEO_DEFAULT_REGION_DESC,
-            .region_desc_len = strlen(KOLISEO_DEFAULT_REGION_DESC),
-            .region_type = KLS_None
-        };
-        kls->hooks.on_push_handler(kls, padding, (void*)&ar_args);
-        */
-        current->hooks.on_push_handler(current, padding, __func__, NULL);
+    for (size_t i=0; i < kls->hooks_len; i++) {
+        if (current->hooks[i].on_push_handler != NULL) {
+            /*
+            struct KLS_EXTENSION_AR_DEFAULT_ARGS {
+                const char* region_name;
+                size_t region_name_len;
+                const char* region_desc;
+                size_t region_desc_len;
+                int region_type;
+            };
+            struct KLS_EXTENSION_AR_DEFAULT_ARGS ar_args = {
+                .region_name = KOLISEO_DEFAULT_REGION_NAME,
+                .region_name_len = strlen(KOLISEO_DEFAULT_REGION_NAME),
+                .region_desc = KOLISEO_DEFAULT_REGION_DESC,
+                .region_desc_len = strlen(KOLISEO_DEFAULT_REGION_DESC),
+                .region_type = KLS_None
+            };
+            kls->hooks.on_push_handler(kls, padding, (void*)&ar_args);
+            */
+            current->hooks[i].on_push_handler(current, padding, __func__, NULL);
+        }
     }
 
     char h_size[200];
@@ -1375,9 +1501,11 @@ void *kls_temp_push_zero_ext_dbg(Koliseo_Temp *t_kls, ptrdiff_t size,
     current->prev_offset = current->offset;
     current->offset += padding + size * count;
 
-    if (current->hooks.on_temp_push_handler != NULL) {
-        // Call on_temp_push extension with empty user arg
-        current->hooks.on_temp_push_handler(t_kls, padding, __func__, NULL);
+    for (size_t i=0; i < kls->hooks_len; i++) {
+        if (current->hooks[i].on_temp_push_handler != NULL) {
+            // Call on_temp_push extension with empty user arg
+            current->hooks[i].on_temp_push_handler(t_kls, padding, __func__, NULL);
+        }
     }
 
     char h_size[200];
@@ -1568,9 +1696,11 @@ void kls_free(Koliseo *kls)
     while (current) {
         Koliseo* next = current->next;
         current->next = NULL;
-        if (current->hooks.on_free_handler != NULL) {
-            // Call on_free() extension
-            current->hooks.on_free_handler(current);
+        for (size_t i=0; i < current->hooks_len; i++) {
+            if (current->hooks[i].on_free_handler != NULL) {
+                // Call on_free() extension
+                current->hooks[i].on_free_handler(current);
+            }
         }
         if (current->has_temp == 1) {
 #ifdef KLS_DEBUG_CORE
@@ -1663,9 +1793,11 @@ Koliseo_Temp *kls_temp_start_dbg(Koliseo *kls, Koliseo_Loc loc)
 
     current->has_temp = 1;
     current->t_kls = tmp;
-    if (current->hooks.on_temp_start_handler != NULL) {
-        // Call on_temp_start extension
-        current->hooks.on_temp_start_handler(tmp);
+    for (size_t i=0; i < kls->hooks_len; i++) {
+        if (current->hooks[i].on_temp_start_handler != NULL) {
+            // Call on_temp_start extension
+            current->hooks[i].on_temp_start_handler(tmp);
+        }
     }
 #ifdef KLS_DEBUG_CORE
     kls_log(current, "KLS", "Prepared new Temp KLS.");
@@ -1692,9 +1824,11 @@ void kls_temp_end(Koliseo_Temp *tmp_kls)
         exit(EXIT_FAILURE);
     }
 
-    if (kls_ref->hooks.on_temp_free_handler != NULL) {
-        // Call on_temp_free() extension
-        kls_ref->hooks.on_temp_free_handler(tmp_kls);
+    for (size_t i=0; i < kls_ref->hooks_len; i++) {
+        if (kls_ref->hooks[i].on_temp_free_handler != NULL) {
+            // Call on_temp_free() extension
+            kls_ref->hooks[i].on_temp_free_handler(tmp_kls);
+        }
     }
 
 #ifdef KLS_DEBUG_CORE