+++ tests/error/oom.c
@@ -11,21 +11,13 @@ int main(void)
 {
     //Init the arena
     Koliseo* kls = kls_new(1000);
-    kls->conf.kls_growable = 1;
-    Koliseo_Temp* kls_t = kls_temp_start(kls);
 
     //Use the arena (see demo for Koliseo_Temp usage)
     Example* e = NULL;
     for(int i = 0; i < 500; i++) {
-        e = KLS_PUSH_T(kls_t,Example);
-    }
-    e->val = 42;
-
-    kls_temp_end(kls_t);
-
-    for(int i = 0; i < 500; i++) {
         e = KLS_PUSH(kls,Example);
     }
+    e->val = 42;
 
     //Free the arena
     kls_free(kls);