+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## [0.5.3] - Unreleased
+
+### Added
+
+- Add `KLS_MAX_EXTENSIONS` and permit more than one extension
+- Add `KLS_DEFAULT_EXTENSIONS_LEN` to permit silent upgrade of `kls_new()` variants not ending in `_ext`
+- Add `oom`, `oom_temp` tests
+
+### Changed
+
+- Fix: check for non-NULL `OOM_handler `when `OOM_happened`
+- Fix experimental `kls_list_pop()` in `kls_region.c`
+- Made all `kls_new()` variants into proper functions
+- Fix demo in region mode
+- Bump `invil` to `0.2.26`
+- Bump `amboso` to `2.0.12`
+
 ## [0.5.2] - 2025-08-01
 
 ### Added